+++ uui/components/forms/Form.tsx
@@ -1,10 +1,8 @@
 import * as React from 'react';
-import {
-    Form as UuiForm, FormProps, useUuiContext, INotification,
-} from '@epam/uui-core';
-import { Text, RichTextView } from '../typography';
-import { ConfirmationModal, WarningNotification } from '../overlays';
+import { Form as UuiForm, FormProps, useUuiContext } from '@epam/uui-core';
+import { ConfirmationModal } from '../overlays';
 import { i18n } from '../../i18n';
+import { showUnsavedChangesNotification } from './showUnsavedChangesNotification';
 
 export function Form<T>(props: FormProps<T>) {
     const context = useUuiContext();
@@ -13,26 +11,8 @@ export function Form<T>(props: FormProps<T>) {
         return context.uuiModals.show<boolean>((modalProps) => <ConfirmationModal caption={ i18n.form.modals.beforeLeaveMessage } { ...modalProps } />);
     }, [context.uuiModals]);
 
-    const loadUnsavedChanges = (): Promise<void> => {
-        return context.uuiNotifications
-            .show(
-                (props: INotification) => (
-                    <WarningNotification
-                        { ...props }
-                        actions={ [
-                            {
-                                name: i18n.form.notifications.actionButtonCaption,
-                                action: props.onSuccess,
-                            },
-                        ] }
-                    >
-                        <RichTextView>
-                            <Text>{i18n.form.notifications.unsavedChangesMessage}</Text>
-                        </RichTextView>
-                    </WarningNotification>
-                ),
-                { duration: 5, position: 'bot-left' },
-            );
+    const loadUnsavedChanges = (): Promise<void | boolean> => {
+        return showUnsavedChangesNotification(context.uuiNotifications);
     };
 
     return <UuiForm<T> loadUnsavedChanges={ loadUnsavedChanges } beforeLeave={ beforeLeave } { ...props } />;