+++ uui/components/forms/useForm.tsx
@@ -1,11 +1,8 @@
 import React, { useCallback } from 'react';
-import {
-    useUuiContext, UseFormProps, useForm as uuiUseForm, INotification,
-} from '@epam/uui-core';
-import { WarningNotification } from '../overlays';
-import { Text, RichTextView } from '../typography';
+import { useUuiContext, UseFormProps, useForm as uuiUseForm } from '@epam/uui-core';
 import { ConfirmationModal } from '../overlays/ConfirmationModal';
 import { i18n } from '../../i18n';
+import { showUnsavedChangesNotification } from './showUnsavedChangesNotification';
 
 export function useForm<T>(props: UseFormProps<T>) {
     const context = useUuiContext();
@@ -14,25 +11,8 @@ export function useForm<T>(props: UseFormProps<T>) {
         return context.uuiModals.show<boolean>((modalProps) => <ConfirmationModal caption={ i18n.form.modals.beforeLeaveMessage } { ...modalProps } />);
     }, [context.uuiModals]);
 
-    const loadUnsavedChanges = (): Promise<void> => {
-        return context.uuiNotifications.show(
-            (props: INotification) => (
-                <WarningNotification
-                    { ...props }
-                    actions={ [
-                        {
-                            name: i18n.form.notifications.actionButtonCaption,
-                            action: props.onSuccess,
-                        },
-                    ] }
-                >
-                    <RichTextView>
-                        <Text>{i18n.form.notifications.unsavedChangesMessage}</Text>
-                    </RichTextView>
-                </WarningNotification>
-            ),
-            { duration: 5, position: 'bot-left' },
-        );
+    const loadUnsavedChanges = (): Promise<void | boolean> => {
+        return showUnsavedChangesNotification(context.uuiNotifications);
     };
 
     return uuiUseForm({ beforeLeave, loadUnsavedChanges, ...props });